% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_weights_reference.R
\docType{data}
\name{gene_weights_reference}
\alias{gene_weights_reference}
\title{Gene Expression Values for Pancreatic Cancer Cell Lines exposed to Hypoxia}
\format{A data frame with columns:
\describe{
 \item{gene_weight}{Float. Gene weighting learned from reference data.}
 \item{gene_ids}{String. The ensembl id of the gene.}
}}
\source{
Derived data
}
\usage{
data(gene_weights_reference)
}
\description{
Gene Expression Values for Pancreatic Cancer Cell Lines exposed to Hypoxia
}
\examples{
\dontrun{
 gene_weights_reference
}
}
\keyword{datasets}
