% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_samp_df.R
\docType{data}
\name{new_samp_df}
\alias{new_samp_df}
\title{Gene Expression Values for PDAC Cancer Cell Lines exposed to Hypoxia}
\format{An object of class \code{data.frame} with 12 rows and 27 columns.}
\source{
Generated by Philipp Markolin, files will be uploaded on GEO
}
\usage{
data(new_samp_df)
}
\description{
A data frame with columns:
\describe{
 \item{sample_id}{String. The name of the sample. Samples with "hyp"
 or "norm" in the sample id are cell lines that were exposed to
 hypoxic or normoxic conditions respectively. Samples with "ctrl" or
 "noHIF" were samples that were able to produce a HIF-mediated hypoxic
 response or not, respectively.}
 \item{ENSG00000074410}{Int. Gene expression value for this gene.}
}
}
\examples{
\dontrun{
 new_samp_df
}
}
\keyword{datasets}
