% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_expr_df.R
\docType{data}
\name{tcga_expr_df}
\alias{tcga_expr_df}
\title{Gene Expression Values for PDAC Cancer Cell Lines exposed to Hypoxia}
\format{An object of class \code{data.frame} with 9461 rows and 54 columns.}
\source{
This data is generated by the TCGA Research Network:
https://www.cancer.gov/tcga and downloaded from the NCI Genomic Data Commons.
}
\usage{
data(tcga_expr_df)
}
\description{
A data frame with columns:
\describe{
 \item{tcga_id}{String. TCGA aliquot barcode}
 \item{study}{String. TCGA study abbreviation}
 \item{is_normal}{Boolean. TRUE if sample is adjacent normal,
 FALSE if tumor.}
 \item{libsize_75percent}{Float. Library size as estimated by the 75th
 quartile.}
 \item{ENSG00000070831}{String. Library size normalized gene expression value for
 this gene.}
}
}
\examples{
\dontrun{
 tcga_expr_df
}
}
\keyword{datasets}
